refactor(SponsorsUI): split SponsorsView into a loader and a content view
SponsorsContentView draws from a ScreenState it is given, so the three
appearances no longer depend on a fetch. The enum replaces an isLoading flag
that could contradict the sponsor list.
