+++ Packages/SponsorsUI/Sources/SponsorsUI/FilterChip.swift
@@ -1,0 +1,23 @@
+import SharedAssets
+import SwiftUI
+
+struct FilterChip: View {
+    let title: String
+    let isSelected: Bool
+    let action: () -> Void
+
+    var body: some View {
+        Button(action: action) {
+            Text(title)
+                .font(.subheadline.weight(isSelected ? .semibold : .regular))
+                .foregroundColor(isSelected ? .white : .primary)
+                .padding(.horizontal, 16)
+                .padding(.vertical, 8)
+                .background(
+                    Capsule()
+                        .fill(isSelected ? Color.accent : Color.secondary.opacity(0.1))
+                )
+        }
+        .buttonStyle(PlainButtonStyle())
+    }
+}
