+++ Packages/SponsorsUI/Sources/SponsorsUI/SponsorsView.swift
@@ -1,270 +1,25 @@
-import DesignKit
-import SharedAssets
 import SponsorsFeature
 import SwiftUI
-import UIComponents
 
+/// The sponsors screen. Fetches the sponsors, then hands them to the view that draws them.
 public struct SponsorsView: View {
     @StateObject private var viewModel = SponsorsViewModel()
-    @State private var isLoading = true
-    @State private var selectedSponsorLevel: SponsorLevel?
+    @State private var state = SponsorsContentView.ScreenState.loading
 
     public init() {}
 
     public var body: some View {
-        ScrollView {
-            VStack(spacing: 0) {
-                FancyHeaderView(
-                    title: "Sponsors",
-                    foregroundImage: Image.swiftLeedsIcon
-                )
-
-                if isLoading {
-                    loadingView
-                        .padding(.top, Padding.screen)
-                } else if viewModel.sponsors.isEmpty {
-                    emptyStateView
-                        .padding(.top, Padding.screen)
-                } else {
-                    sponsorsList
-                }
+        SponsorsContentView(state: state)
+            .task {
+                await loadSponsors()
             }
-        }
-        .background(Color.background, ignoresSafeAreaEdges: .all)
-        .edgesIgnoringSafeArea(.top)
-        .scrollIndicators(.hidden)
-        .task {
-            await loadSponsors()
-        }
-    }
-
-    private var loadingView: some View {
-        VStack(spacing: Padding.cellGap) {
-            ProgressView()
-                .scaleEffect(1.2)
-                .tint(.accent)
-            Text("Loading sponsors...")
-                .font(.subheadline)
-                .foregroundColor(.secondary)
-        }
-        .frame(maxWidth: .infinity, minHeight: 300)
-        .padding()
-    }
-
-    private var emptyStateView: some View {
-        VStack(spacing: Padding.cellGap) {
-            Image(systemName: "sparkles")
-                .font(.system(size: 50))
-                .foregroundColor(.secondary)
-            Text("No sponsors available")
-                .font(.headline)
-            Text("Check back later for updates")
-                .font(.subheadline)
-                .foregroundColor(.secondary)
-        }
-        .frame(maxWidth: .infinity, minHeight: 300)
-        .padding()
-    }
-
-    private var tierEmptyStateView: some View {
-        VStack(spacing: Padding.cellGap) {
-            if let selectedLevel = selectedSponsorLevel {
-                Image(systemName: iconForLevel(selectedLevel))
-                    .font(.system(size: 50))
-                    .foregroundColor(.secondary.opacity(0.6))
-
-                Text("No \(selectedLevel.rawValue.capitalized) Sponsors")
-                    .font(.headline)
-                    .foregroundColor(.primary)
-
-                Text("This tier doesn't have any sponsors yet")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-                    .multilineTextAlignment(.center)
-            }
-        }
-        .frame(maxWidth: .infinity, minHeight: 200)
-        .padding()
-    }
-
-    private var sponsorsList: some View {
-        VStack(spacing: Padding.cellGap) {
-            filterChips
-
-            if displayedLevels.isEmpty && selectedSponsorLevel != nil {
-                tierEmptyStateView
-            } else {
-                ForEach(displayedLevels, id: \.self) { level in
-                    sectionView(for: level)
-                }
-            }
-        }
-        .padding(.horizontal, horizontalPadding)
-        .padding(.bottom, Padding.cellGap)
-    }
-
-    private var filterChips: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: 8) {
-                FilterChip(
-                    title: "All",
-                    isSelected: selectedSponsorLevel == nil,
-                    action: { selectedSponsorLevel = nil }
-                )
-
-                ForEach(SponsorLevel.allCases, id: \.self) { level in
-                    FilterChip(
-                        title: level.rawValue.capitalized,
-                        isSelected: selectedSponsorLevel == level,
-                        action: {
-                            withAnimation(.spring(response: 0.3, dampingFraction: 0.7)) {
-                                selectedSponsorLevel = level == selectedSponsorLevel ? nil : level
-                            }
-                        }
-                    )
-                }
-            }
-        }
-        .padding(.vertical, 8)
-    }
-
-    private var displayedLevels: [SponsorLevel] {
-        if let selectedLevel = selectedSponsorLevel {
-            return viewModel.sponsors.rankedLevels.filter { $0 == selectedLevel }
-        }
-        return viewModel.sponsors.rankedLevels
-    }
-
-    private var gridColumns: [GridItem] {
-        return Array(repeating: GridItem(.flexible(), spacing: Padding.cellGap), count: columnCount)
-    }
-
-    private var columnCount: Int {
-        #if os(iOS)
-        return UIDevice.current.userInterfaceIdiom == .pad ? 3 : 2
-        #else
-        return 2
-        #endif
-    }
-
-    private var horizontalPadding: CGFloat {
-        #if os(iOS)
-        return UIDevice.current.userInterfaceIdiom == .pad ? Padding.screen * 2 : Padding.screen
-        #else
-        return Padding.screen
-        #endif
-    }
-
-    private func sectionView(for level: SponsorLevel) -> some View {
-        VStack(alignment: .leading, spacing: Padding.stackGap) {
-            sectionHeader(for: level)
-
-            switch level {
-            case .platinum:
-                VStack(spacing: Padding.cellGap) {
-                    ForEach(viewModel.sponsors.sponsors(at: level)) { sponsor in
-                        SponsorTileView(sponsor: sponsor)
-                            .transition(.asymmetric(
-                                insertion: .scale.combined(with: .opacity),
-                                removal: .scale.combined(with: .opacity)
-                            ))
-                    }
-                }
-            case .gold, .silver:
-                LazyVGrid(
-                    columns: gridColumns,
-                    alignment: .leading,
-                    spacing: Padding.cellGap
-                ) {
-                    ForEach(viewModel.sponsors.sponsors(at: level)) { sponsor in
-                        SponsorTileView(sponsor: sponsor)
-                            .transition(.asymmetric(
-                                insertion: .scale.combined(with: .opacity),
-                                removal: .scale.combined(with: .opacity)
-                            ))
-                    }
-                }
-            }
-        }
-        .padding(.bottom, Padding.cellGap)
-    }
-
-    private func sectionHeader(for sponsorLevel: SponsorLevel) -> some View {
-        HStack {
-            Image(systemName: iconForLevel(sponsorLevel))
-                .font(.caption)
-                .foregroundColor(colorForLevel(sponsorLevel))
-
-            Text("\(sponsorLevel.rawValue.capitalized) Sponsors")
-                .font(.headline.weight(.semibold))
-                .foregroundColor(.primary)
-
-            Spacer()
-
-            Text("\(sponsorCount(for: sponsorLevel))")
-                .font(.caption)
-                .foregroundColor(.secondary)
-                .padding(.horizontal, 8)
-                .padding(.vertical, 2)
-                .background(Color.secondary.opacity(0.1))
-                .clipShape(Capsule())
-        }
-        .accessibilityAddTraits(.isHeader)
-        .padding(.vertical, 8)
-    }
-
-    private func iconForLevel(_ level: SponsorLevel) -> String {
-        switch level {
-        case .platinum: return "crown.fill"
-        case .gold: return "star.fill"
-        case .silver: return "star"
-        }
-    }
-
-    private func colorForLevel(_ level: SponsorLevel) -> Color {
-        switch level {
-        case .platinum: return .purple
-        case .gold: return .yellow
-        case .silver: return .gray
-        }
-    }
-
-    private func sponsorCount(for level: SponsorLevel) -> Int {
-        viewModel.sponsors.sponsors(at: level).count
     }
 
     private func loadSponsors() async {
-        do {
-            try await viewModel.loadSponsors()
-            withAnimation(.easeInOut(duration: 0.3)) {
-                isLoading = false
-            }
-        } catch {
-            withAnimation(.easeInOut(duration: 0.3)) {
-                isLoading = false
-            }
+        try? await viewModel.loadSponsors()
+        withAnimation(.easeInOut(duration: 0.3)) {
+            state = .init(viewModel.sponsors)
         }
-    }
-}
-
-struct FilterChip: View {
-    let title: String
-    let isSelected: Bool
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            Text(title)
-                .font(.subheadline.weight(isSelected ? .semibold : .regular))
-                .foregroundColor(isSelected ? .white : .primary)
-                .padding(.horizontal, 16)
-                .padding(.vertical, 8)
-                .background(
-                    Capsule()
-                        .fill(isSelected ? Color.accent : Color.secondary.opacity(0.1))
-                )
-        }
-        .buttonStyle(PlainButtonStyle())
     }
 }
